ITrace: ProxyStub generation support added
