+++ Source/protocols/CMakeLists.txt
@@ -20,6 +20,7 @@ set(TARGET ${NAMESPACE}Protocols)
 find_package(ZLIB REQUIRED)
 
 ProxyStubGenerator(NAMESPACE "WPEFramework::RPC" INPUT "${CMAKE_CURRENT_SOURCE_DIR}/../com/ICOM.h" OUTDIR "${CMAKE_CURRENT_BINARY_DIR}/generated")
+ProxyStubGenerator(NAMESPACE "WPEFramework::Trace" INPUT "${CMAKE_CURRENT_SOURCE_DIR}/../com/ITrace.h" OUTDIR "${CMAKE_CURRENT_BINARY_DIR}/generated")
 
 # Take care the sources are listed !!! TWICE !!! in this setup. raised a JIRA ticket to solve this for now, also scroll down and list the 
 # new source file again. This is in case CMAKE forgets :-)
@@ -31,11 +32,11 @@ set(SOURCES_COM
         com/Ids.h
         com/ICOM.h
         com/IUnknown.h
+        com/ITrace.h
         com/Messages.h
         com/Administrator.cpp
         com/Communicator.cpp
         com/ITracing.cpp
-        com/ProxyStubs.cpp.in
         com/IValueIterator.h
         com/IUnknown.cpp
         com/IStringIterator.h
@@ -78,6 +79,7 @@ set(COM_INCLUDES
         "${CMAKE_CURRENT_BINARY_DIR}/com/Ids.h"
         "${CMAKE_CURRENT_BINARY_DIR}/com/ICOM.h"
         "${CMAKE_CURRENT_BINARY_DIR}/com/IUnknown.h"
+        "${CMAKE_CURRENT_BINARY_DIR}/com/ITrace.h"
         "${CMAKE_CURRENT_BINARY_DIR}/com/Messages.h"
         "${CMAKE_CURRENT_BINARY_DIR}/com/IRPCIterator.h"
         "${CMAKE_CURRENT_BINARY_DIR}/com/IStringIterator.h"
@@ -104,19 +106,14 @@ set(PUBLIC_HEADERS
         Module.h
         )
 
-set(PROXYSTUB_COM_SOURCE "${CMAKE_CURRENT_BINARY_DIR}/generated/ProxyStubs_COM.cpp")
-
-configure_file( "${CMAKE_CURRENT_BINARY_DIR}/com/ProxyStubs.cpp.in"
-                "${CMAKE_CURRENT_BINARY_DIR}/com/ProxyStubs.cpp"
-                @ONLY)
-
 # Construct a library object
 add_library(${TARGET} SHARED
         Module.cpp
+        "${CMAKE_CURRENT_BINARY_DIR}/generated/ProxyStubs_COM.cpp"
+        "${CMAKE_CURRENT_BINARY_DIR}/generated/ProxyStubs_Trace.cpp"
         "${CMAKE_CURRENT_BINARY_DIR}/com/Administrator.cpp"
         "${CMAKE_CURRENT_BINARY_DIR}/com/Communicator.cpp"
         "${CMAKE_CURRENT_BINARY_DIR}/com/ITracing.cpp"
-        "${CMAKE_CURRENT_BINARY_DIR}/com/ProxyStubs.cpp"
         "${CMAKE_CURRENT_BINARY_DIR}/com/IUnknown.cpp"
         "${CMAKE_CURRENT_BINARY_DIR}/websocket/URL.cpp"
         "${CMAKE_CURRENT_BINARY_DIR}/websocket/JSONWebToken.cpp"
